Split session actor isolation files for CI
