+++ crates/tandem-server/src/http/tests/sessions.rs
@@ -3,3 +3,4 @@ include!("sessions_parts/part02.rs");
 include!("sessions_parts/part03.rs");
 include!("sessions_parts/part04.rs");
 include!("sessions_parts/part05.rs");
+include!("sessions_parts/part06.rs");
